update5

+++ src/main/java/com/example/product/service/ProductService.java
@@ -4,6 +4,7 @@ package com.example.product.service;
 import com.example.product.entitiy.Product;
 import com.example.product.repository.ProductRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -32,6 +33,9 @@ public class ProductService {
     }
 
     public void deleteProduct(Long id) {
+        if (!productRepo.existsById(id)) {
+            throw new RuntimeException("CONTENT NOT FOUND");
+        }
         productRepo.deleteById(id);
     }
